requirements.txt
logbook
requests
loguru
python-nmap
ConfigParser